requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
spacy==3.7.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
python-dateutil==2.8.2